+++ app/src/main/java/org/example/model/filters/filterModels/Parameter.java
@@ -7,13 +7,13 @@ import java.util.List;
 public class Parameter {
     public FieldType type;
     public Object parameter;
-    public Integer max;
-    public Integer min;
+    public Double max;
+    public Double min;
     public Integer step;
     public int index;
     public List<String> link;
 
-    public Parameter(FieldType type, Object parameter, Integer max, Integer min, Integer step, List<String> link) {
+    public Parameter(FieldType type, Object parameter, Double max, Double min, Integer step, List<String> link) {
         this.type = type;
         this.parameter = parameter;
         this.max = max;
